server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "pydantic-ai-slim[groq]>=1.67.0",
    "pydantic-ai[examples,groq]>=1.67.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "supabase>=2.28.3",
    "supabase-auth>=2.28.3",
    "uvicorn>=0.41.0",
]
